Refs #4270. MinGW compile fixes for mythmusic.
git-svn-id: http://svn.mythtv.org/svn/trunk@15123 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythplugins/mythmusic/mythmusic/mainvisual.cpp
@@ -149,7 +149,7 @@ void MainVisual::setVisual(const QString &name)
     {
         if (pVisFactory->name() == visName)
         {
-            vis = pVisFactory->create(this, winId(), pluginName);
+            vis = pVisFactory->create(this, (long int) winId(), pluginName);
             vis->resize(size());
             fps = vis->getDesiredFPS();
             break;
